fix(scripts): CF-1 schema_version int64 cast in migrate_instruments_store_v9 + regression test
When the existing _index parquet stores schema_version as object/string dtype, assigning
MANIFEST_SCHEMA_VERSION=9 preserves object dtype and pyarrow may serialise as str '9' not
int64 9. Add explicit .astype('int64') cast so E4 VM apply writes integer 9 (CF-1 GREEN).
Add regression test: inputs schema_version='9' (string), asserts dtype==int64, value==9.


Quickmerge: agent

+++ scripts/migrate_instruments_store_v9.py
@@ -329,8 +329,10 @@ def transform_index_v9(df: pd.DataFrame, asset_group: str) -> tuple[pd.DataFrame
     # columns for the sports owner (sports_manifest_canonicalisation_2026_06_01.md).
     is_sports = asset_group == "sports"
 
-    # CF-1 schema_version
+    # CF-1 schema_version — explicit int64 cast so existing string-typed parquet columns
+    # (object dtype from historical writes) don't propagate as str '9' after to_parquet.
     out["schema_version"] = MANIFEST_SCHEMA_VERSION
+    out["schema_version"] = out["schema_version"].astype("int64")
     # CF-2 asset_group on rows
     blank_ag = out["asset_group"].str.len() == 0
     out.loc[blank_ag, "asset_group"] = asset_group
